Patchset(2/4) - EzimDtxSemantics Attributes to public
In order for the EzimReceiver class to function properly we had to
change EzminDtxSemantics into a data class. To accomplish this we had to
make all of the attributes of the class public so that they could be
accessed by other classes within the system. This is necessary because
the method calls of the EzimReceiver class required access to the
attributes of EzimDtxSemantics.