Ezim New Class "EzimSocketBinder" Implements Runnable . New Functions/Variables Added
The new "EzimSocketBinder" class now implements the Runnable interface.

A new function "EzimFileType" has been added that will take a string as
a parameter which will define whether it is a file request , confirm ,
response or file sending.

The variables associated with the functions has been added.

A new function "doRun" has been created that will be called from the 4
other classes. It takes a unique string as a parameter that will depend
on which class is calling it.

The "run" function for this class has been created.