Class "EzimSocketBinder" run() function Implemented. The classes to be merged were Modified accordingly.
The "run" function for the class "EzimSocketBinder" has been created .
It will combine the 4 classes that are to be merged and process their
operations accordingly , depending on which class called the "doRun"
function which will then call this one.

The "run" function for the 4 classes mentioned in patchset 0 has been
modified with a single "doRun" function that will make a call to the new
class "EzimSocketBinder" . It will have a unique string as parameter
depending
on which class makes the call.