EzimDtxTakerThread now refers to EzimReceiver and not EzimDtxSemantics
Since all the changes are now complete for the EzimReceiver class to
function on its own, in this patch set we focused on changing the method
calls in other classes to now call the appropriate methods in
EzimReceiver. In this case there was only one call to the parser method
that was in EzimDtxSemantics which we now switched to a parser method
call in EzimReciever.